+++ spec/flash_spec.rb
@@ -1,67 +1,50 @@
 # frozen_string_literal: true
 
 RSpec.describe GovukPersonalisation::Flash do
-  describe "#decode_session" do
-    let(:encoded_session) { "" }
-    subject(:decoded) { GovukPersonalisation::Flash.decode_session(encoded_session) }
+  describe "#decode" do
+    let(:encoded) { "" }
+    subject(:decoded) { GovukPersonalisation::Flash.decode(encoded) }
 
     it "returns nil" do
       expect(decoded).to be_nil
     end
 
-    context "the encoded session has one part" do
-      let(:encoded_session) { "session-id" }
+    context "there are encoded messages" do
+      let(:encoded) { messages.join(",") }
+      let(:messages) { %w[foo bar baz] }
 
-      it "returns the session and no flash messages" do
-        expect(decoded).to eq([encoded_session, []])
-      end
-    end
-
-    context "the encoded session has two parts" do
-      let(:encoded_session) { "#{session_id}$$#{flash.join(',')}" }
-      let(:session_id) { "session-id" }
-      let(:flash) { %w[foo bar baz] }
-
-      it "returns the session and the flash messages" do
-        expect(decoded).to eq([session_id, flash])
-      end
-
-      context "the flash part is empty" do
-        let(:flash) { [] }
-
-        it "returns the session and no flash messages" do
-          expect(decoded).to eq([session_id, []])
-        end
+      it "returns a hash of messages" do
+        expect(decoded).to eq(messages.index_with { |_| true })
       end
     end
   end
 
-  describe "#encode_session" do
-    let(:session) { "session-id" }
-    let(:flash) { %w[foo bar baz] }
-    subject(:encoded) { GovukPersonalisation::Flash.encode_session(session, flash) }
+  describe "#encode" do
+    let(:messages) { %w[foo bar baz] }
+    let(:flash) { messages.index_with { |_| true } }
+    subject(:encoded) { GovukPersonalisation::Flash.encode(flash) }
 
-    it "returns a two-part encoded session" do
-      expect(encoded).to eq("#{session}$$#{flash.join(',')}")
+    it "encodes the messages" do
+      expect(encoded).to eq(messages.join(","))
     end
 
-    it "round-trips through #decode_session" do
-      expect(GovukPersonalisation::Flash.decode_session(encoded)).to eq([session, flash])
+    it "round-trips through #decode" do
+      expect(GovukPersonalisation::Flash.decode(encoded)).to eq(flash)
     end
 
     context "the flash is empty" do
       let(:flash) { [] }
 
-      it "returns a one-part encoded session" do
-        expect(encoded).to eq(session)
+      it "returns nil" do
+        expect(encoded).to eq(nil)
       end
     end
 
     context "the flash is nil" do
       let(:flash) { nil }
 
-      it "returns a one-part encoded session" do
-        expect(encoded).to eq(session)
+      it "returns nil" do
+        expect(encoded).to eq(nil)
       end
     end
   end