+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- BREAKING: Move session encoding / decoding out of `GovukPersonalisation::Flash` and into `GovukPersonalisation::Session`
+
 # 0.6.0
 
 - Add `GovukPersonalisation::Flash` and helper methods to the concern ([#9](https://github.com/alphagov/govuk_personalisation/pull/9))